+++ src/components/index.js
@@ -1,9 +1,9 @@
 import { AppNavigation } from "./AppNavigation";
-import { ButtonCta } from "./ButtonCta";
+import { ButtonCta } from "./utilities/ButtonCta";
 import { Header } from "./Header";
-import { PageContainer } from "./PageContainer";
+import { PageContainer } from "./utilities/PageContainer";
 import { NotesdModal } from "./NotesModal";
-import { EmptyState } from './EmptyState';
-import { PageHeader } from './PageHeader';
+import { EmptyState } from './utilities/EmptyState';
+import { PageHeader } from './utilities/PageHeader';
 import { NotesContainer, NoteCard } from './NoteCard'
 export {AppNavigation,ButtonCta,Header,PageContainer,NotesdModal,EmptyState,PageHeader,NotesContainer,NoteCard}
