+++ src/components/utilities/HeadingLarge.jsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const HeadingLarge = styled.h1`
+margin-right: auto;
+`;
